# lib/api-zod/src/generated/types/adminAdvogado.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminAdvogadoPaymentStatus } from './adminAdvogadoPaymentStatus';
import type { AdminAdvogadoSituacaoOab } from './adminAdvogadoSituacaoOab';

export interface AdminAdvogado {
  id: number;
  nome: string;
  oab: string;
  email: string;
  createdAt?: string | null;
  paymentStatus?: AdminAdvogadoPaymentStatus;
  adminAtivo: boolean;
  oabVerificada: boolean;
  situacaoOab?: AdminAdvogadoSituacaoOab;
}