# lib/api-zod/src/generated/types/adminAdvogadoDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminAdvogadoDetailPaymentStatus } from './adminAdvogadoDetailPaymentStatus';
import type { AdminAdvogadoDetailSituacaoOab } from './adminAdvogadoDetailSituacaoOab';
import type { AtividadeLog } from './atividadeLog';
import type { Cidade } from './cidade';

export interface AdminAdvogadoDetail {
  id: number;
  nome: string;
  oab: string;
  email: string;
  cpf?: string | null;
  createdAt?: string | null;
  paymentStatus?: AdminAdvogadoDetailPaymentStatus;
  plano?: string | null;
  adminAtivo: boolean;
  oabVerificada: boolean;
  situacaoOab?: AdminAdvogadoDetailSituacaoOab;
  areas: string[];
  cidades: Cidade[];
  atendeOnline: boolean;
  atividades: AtividadeLog[];
}