# lib/api-zod/src/generated/types/atividadeLog.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AtividadeLog {
  id: number;
  acao: string;
  feitoPor: string;
  data: string;
}