# lib/api-zod/src/generated/types/updateAdminAdvogadoInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAdminAdvogadoInputSituacaoOab } from './updateAdminAdvogadoInputSituacaoOab';

export interface UpdateAdminAdvogadoInput {
  adminAtivo?: boolean;
  oabVerificada?: boolean;
  situacaoOab?: UpdateAdminAdvogadoInputSituacaoOab;
}